test(scene3d): consolidate hub remount coverage
