+++ hub/scene3d/roundtrip_test.go
@@ -1,12 +1,16 @@
 package scene3d
 
 import (
+	"bytes"
 	"encoding/json"
+	"errors"
 	"fmt"
+	"reflect"
 	"testing"
 
 	"m31labs.dev/gosx/crdt"
 	"m31labs.dev/gosx/scene"
+	"m31labs.dev/gosx/scene/capability"
 )
 
 // allCommandKinds lists every kind scene/diff.go declares. The list is written
@@ -312,6 +316,196 @@ func TestApplyValidatesBeforeWriting(t *testing.T) {
 	}
 	if len(commands) != 0 {
 		t.Fatalf("partially applied stream wrote %d commands", len(commands))
+	}
+}
+
+// TestDiffSceneRejectsRemountOnlyBeforeMutation proves a remount-only diff is
+// refused before it can create history or wake a watcher.
+func TestDiffSceneRejectsRemountOnlyBeforeMutation(t *testing.T) {
+	doc, bound := newBoundDoc(t)
+	var watched int
+	bound.Watch(func(commands []scene.Command) { watched += len(commands) })
+
+	before, err := doc.Save()
+	if err != nil {
+		t.Fatal(err)
+	}
+	commands, hash, err := bound.DiffScene(scene.SceneIR{}, scene.SceneIR{
+		Schema: "gosx.scene3d.ir.v2",
+	}, "remount-only")
+	if err == nil {
+		t.Fatal("remount-only diff returned no error")
+	}
+	if commands != nil {
+		t.Fatalf("remount-only diff commands = %v, want nil", commands)
+	}
+	if hash != (crdt.ChangeHash{}) {
+		t.Fatalf("remount-only hash = %s, want zero", hash)
+	}
+	assertRemountError(t, err, []string{"Schema"})
+
+	after, err := doc.Save()
+	if err != nil {
+		t.Fatal(err)
+	}
+	if !bytes.Equal(before, after) {
+		t.Fatal("remount-only diff changed the document snapshot")
+	}
+	if watched != 0 {
+		t.Fatalf("remount-only diff notified watchers %d times", watched)
+	}
+}
+
+// TestDiffSceneRejectsMixedDiffAtomically proves a supported command cannot be
+// committed when the same diff also needs a full scene remount.
+func TestDiffSceneRejectsMixedDiffAtomically(t *testing.T) {
+	doc, bound := newBoundDoc(t)
+	base := scene.SceneIR{Objects: []scene.ObjectIR{{ID: "cube", Kind: "box"}}}
+	if _, _, err := bound.DiffScene(scene.SceneIR{}, base, "seed"); err != nil {
+		t.Fatal(err)
+	}
+
+	var watched int
+	bound.Watch(func(commands []scene.Command) { watched += len(commands) })
+	watched = 0
+	before, err := doc.Save()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	next := base
+	next.Objects = append([]scene.ObjectIR(nil), base.Objects...)
+	next.Objects[0].X = 7
+	next.Schema = "gosx.scene3d.ir.v2"
+	commands, hash, err := bound.DiffScene(base, next, "mixed")
+	if err == nil {
+		t.Fatal("mixed diff returned no error")
+	}
+	if commands != nil {
+		t.Fatalf("mixed diff commands = %v, want nil on rejection", commands)
+	}
+	if hash != (crdt.ChangeHash{}) {
+		t.Fatalf("mixed diff hash = %s, want zero", hash)
+	}
+	assertRemountError(t, err, []string{"Schema"})
+
+	after, err := doc.Save()
+	if err != nil {
+		t.Fatal(err)
+	}
+	if !bytes.Equal(before, after) {
+		t.Fatal("mixed diff changed the document snapshot")
+	}
+	if watched != 0 {
+		t.Fatalf("mixed diff notified watchers %d times", watched)
+	}
+	view, err := bound.View()
+	if err != nil {
+		t.Fatal(err)
+	}
+	if got := objectByID(t, view, "cube").X; got != 0 {
+		t.Fatalf("mixed diff moved cube to %v, want seed value 0", got)
+	}
+}
+
+// TestDiffSceneRemountFieldsAreStable proves every remount field is reported
+// in the sorted order scene.DiffScene promises to its callers.
+func TestDiffSceneRemountFieldsAreStable(t *testing.T) {
+	next := scene.SceneIR{
+		Schema:                "gosx.scene3d.ir.v2",
+		ShadowMaxPixels:       4096,
+		QualityLadder:         []scene.QualityRungIR{{Name: "low"}},
+		QualityStartRung:      2,
+		PointQualityGroups:    map[string]string{"dust": "points"},
+		BackendCaps:           &capability.BackendCaps{Capable: []capability.Backend{capability.BackendWebGL}},
+		ShaderLib:             map[string]string{"sl:0123456789abcdef": "fn main() {}"},
+		MotionProgram:         []byte{1, 2, 3},
+		MaterialMotionProgram: []byte{4, 5, 6},
+	}
+	want := []string{
+		"BackendCaps",
+		"MaterialMotionProgram",
+		"MotionProgram",
+		"PointQualityGroups",
+		"QualityLadder",
+		"QualityStartRung",
+		"Schema",
+		"ShaderLib",
+		"ShadowMaxPixels",
+	}
+
+	_, bound := newBoundDoc(t)
+	for i := 0; i < 2; i++ {
+		_, _, err := bound.DiffScene(scene.SceneIR{}, next, "stable")
+		if err == nil {
+			t.Fatal("remount diff returned no error")
+		}
+		assertRemountError(t, err, want)
+	}
+}
+
+// TestDiffScenePreservesOrdinaryAndEmptyBehavior proves a supported diff still
+// commits its command stream, while an empty diff remains a no-op.
+func TestDiffScenePreservesOrdinaryAndEmptyBehavior(t *testing.T) {
+	doc, bound := newBoundDoc(t)
+	var seen []scene.Command
+	bound.Watch(func(commands []scene.Command) { seen = append(seen, commands...) })
+
+	base := scene.SceneIR{Objects: []scene.ObjectIR{{ID: "cube", Kind: "box"}}}
+	if _, _, err := bound.DiffScene(scene.SceneIR{}, base, "seed"); err != nil {
+		t.Fatal(err)
+	}
+	seen = nil
+	next := base
+	next.Objects = append([]scene.ObjectIR(nil), base.Objects...)
+	next.Objects[0].X = 4
+	commands, hash, err := bound.DiffScene(base, next, "patch")
+	if err != nil {
+		t.Fatal(err)
+	}
+	if hash == (crdt.ChangeHash{}) {
+		t.Fatal("ordinary diff returned a zero hash")
+	}
+	assertCommandsEqual(t, "ordinary diff", commands, seen)
+
+	before, err := doc.Save()
+	if err != nil {
+		t.Fatal(err)
+	}
+	seen = nil
+	commands, hash, err = bound.DiffScene(next, next, "empty")
+	if err != nil {
+		t.Fatal(err)
+	}
+	if commands != nil {
+		t.Fatalf("empty diff commands = %v, want nil", commands)
+	}
+	if hash != (crdt.ChangeHash{}) {
+		t.Fatalf("empty diff hash = %s, want zero", hash)
+	}
+	if len(seen) != 0 {
+		t.Fatalf("empty diff notified watchers with %d commands", len(seen))
+	}
+	after, err := doc.Save()
+	if err != nil {
+		t.Fatal(err)
+	}
+	if !bytes.Equal(before, after) {
+		t.Fatal("empty diff changed the document snapshot")
+	}
+}
+
+func assertRemountError(t *testing.T, err error, want []string) {
+	t.Helper()
+	var remountErr *RemountRequiredError
+	if !errors.As(err, &remountErr) {
+		t.Fatalf("error %T = %v, want *RemountRequiredError", err, err)
+	}
+	if !errors.Is(err, ErrRemountRequired) {
+		t.Fatalf("error %v does not wrap ErrRemountRequired", err)
+	}
+	if !reflect.DeepEqual(remountErr.Fields, want) {
+		t.Fatalf("remount fields = %v, want %v", remountErr.Fields, want)
 	}
 }
